+++ src/error.ts
@@ -6,7 +6,6 @@ import {
   sanitizeStatusCode,
 } from "./utils";
 import { hasProp } from "./utils/internal/object";
-import type { Status } from "./types";
 
 /**
  * H3 Runtime Error
@@ -168,9 +167,7 @@ export function sendError(
   if (event.handled) {
     return;
   }
-  const _code = Number.parseInt(
-    h3Error.statusCode as unknown as string,
-  ) as unknown as Status;
+  const _code = Number.parseInt(h3Error.statusCode as unknown as string);
   setResponseStatus(event, _code, h3Error.statusMessage);
   event.node.res.setHeader("content-type", MIMES.json);
   event.node.res.end(JSON.stringify(responseBody, undefined, 2));
