+++ src/utils/cors/handler.ts
@@ -1,6 +1,5 @@
 import { H3Event } from "../../event";
 import { sendNoContent } from "../response";
-import type { Status } from "../../types";
 import {
   resolveCorsOptions,
   appendCorsPreflightHeaders,
@@ -39,7 +38,7 @@ export function handleCors(event: H3Event, options: H3CorsOptions): boolean {
   const _options = resolveCorsOptions(options);
   if (isPreflightRequest(event)) {
     appendCorsPreflightHeaders(event, options);
-    sendNoContent(event, _options.preflight.statusCode as unknown as Status);
+    sendNoContent(event, _options.preflight.statusCode);
     return true;
   }
   appendCorsHeaders(event, options);
