+++ src/utils/request.ts
@@ -1,14 +1,10 @@
 import { getQuery as _getQuery, decode as decodeURI } from "ufo";
 import { createError } from "../error";
-import type {
-  HTTPHeaderName,
-  HTTPMethod,
-  InferEventInput,
-  RequestHeaders,
-} from "../types";
+import type { HTTPMethod, InferEventInput } from "../types";
 import type { H3Event } from "../event";
 import { validateData, ValidateFunction } from "./internal/validate";
 import { getRequestWebStream } from "./body";
+import { HTTPHeaderName, RequestHeaders } from "../types.headers";
 
 /**
  * Get query the params object from the request URL parsed with [unjs/ufo](https://ufo.unjs.io).
